requirements.txt
# Numerical core
numpy==1.26.4
scipy==1.13.0

# PyTorch — CPU only for Phase 1 (agents don't need GPU)
torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu

# Compression algorithms
zstandard==0.22.0
lz4==4.3.3

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Cryptographic commit-reveal (Phase 2 — install now)
cryptography==42.0.8

# Experiment tracking
wandb==0.17.0

# CLI + formatting
tqdm==4.66.4
pyyaml==6.0.1
rich==13.7.1

# Dev tools
pytest==8.2.2
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.7